+++ hasher-matcher-actioner/hmalib/common/actioner_models.py
@@ -9,9 +9,6 @@ import hmalib.common.config as config
 
 from hmalib.models import Label, MatchMessage, BankedSignal
 from hmalib.common.logging import get_logger
-from hmalib.aws_secrets import AWSSecrets
-
-from threatexchange.api import ThreatExchangeAPI
 
 logger = get_logger(__name__)
 
@@ -105,45 +102,6 @@ class WebhookDeleteActionPerformer(WebhookActionPerformer):
         return delete(self.url)
 
 
-@dataclass
-class ReactActionPerformer(ActionPerformer):
-    @property
-    def reaction(self) -> str:
-        raise NotImplementedError
-
-    def perform_action(self, match_message: MatchMessage) -> None:
-        api_key = AWSSecrets.te_api_key()
-        api = ThreatExchangeAPI(api_key)
-
-        indicator_ids = {
-            dataset_match_details.banked_content_id
-            for dataset_match_details in match_message.matching_banked_signals
-            if dataset_match_details.bank_source == "te"
-        }
-
-        descriptor_ids = {
-            descriptor_id["id"]
-            for indicator_id in indicator_ids
-            for descriptor_id in api.get_threat_descriptors_from_indicator(indicator_id)
-        }
-
-        for id in descriptor_ids:
-            api.react_to_threat_descriptor(id, self.reaction)
-            logger.warning("reacted %s to descriptor %s", self.reaction, id)
-
-
-class ReactInReviewActionPerformer(ReactActionPerformer):
-    reaction = "IN_REVIEW"
-
-
-class ReactIngestedActionPerformer(ReactActionPerformer):
-    reaction = "INGESTED"
-
-
-class ReactSawThisTooActionPerformer(ReactActionPerformer):
-    reaction = "SAW_THIS_TOO"
-
-
 def get_all_subclasses_rec(recursive: t.Set[t.Type]) -> t.Set[t.Type]:
     subclasses = {subclass for cls in recursive for subclass in cls.__subclasses__()}
     union = recursive.union(subclasses)
@@ -214,14 +172,15 @@ if __name__ == "__main__":
     match_message = MatchMessage("key", "hash", banked_signals)
 
     configs: t.List[ActionPerformer] = [
-        ReactInReviewActionPerformer(
-            action_label=ActionLabel("ReactInReview"),
+        WebhookDeleteActionPerformer(
+            action_label=ActionLabel("DeleteWebhook"),
+            url="https://webhook.site/ff7ebc37-514a-439e-9a03-46f86989e195",
         ),
-        ReactSawThisTooActionPerformer(
-            action_label=ActionLabel("ReactSawThisToo"),
+        WebhookPutActionPerformer(
+            action_label=ActionLabel("PutWebook"),
+            url="https://webhook.site/ff7ebc37-514a-439e-9a03-46f86989e195",
         ),
     ]
 
-    # This will react to 4 real descriptors
     for action_config in configs:
         action_config.perform_action(match_message)
